requirements.txt
playwright>=1.40.0
requests
pandas
PyGithub
streamlit
plotly